// src/domain/evaluation/company/dto/evaluation-company-get-list-request.enum.ts
export enum EvaluationStatus {
    INCOMPLETE = 'INCOMPLETE',
    COMPLETE = 'COMPLETE',
}

export enum MemberSearchCategory {
    NAME = 'NAME',
    CONTACT = 'CONTACT',
}

export enum TeamSearchCategory {
    TEAM_NAME = 'TEAM_NAME',
    LEADER_NAME = 'LEADER_NAME',
    LEADER_CONTACT = 'LEADER_CONTACT',
}